requirements.txt
emoji
pandas
scikit-learn
nltk
seaborn
matplotlib
shap
tweepy
transformers
accelerate
python-dotenv
flask
requests
facebook-sdk
python-dotenv
streamlit
transformers
torch
scipy